refactor: replace user functions to separate dir and add possibility to save client index using current ws
